music (#42)
I like the legendofkvothe for title screen and rainbows for the gameplay (it's very playful).  There are lots of these at scottbuckley.com They are under a CC-BY license, which means we have to "give credit" which basically means name the author and provide the link.  This could be done either in a credits scene or on the title screen somewhere.  I'm open to other music options too,  just putting this up so we have something.